+++ Findfs.cmake
@@ -1,3 +1,0 @@
-set(FS_LIBRARIES "stdc++fs")
-
-mark_as_advanced(FS_LIBRARIES)