+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Remmoved
+
+* Findfs.cmake
+
 ## 0.5.0 &ndash; 2020-12-04
 
 ## 0.4.5 &ndash; 2020-11-09